reorganize code to follow conventional repo layout
The main modules for both binaries go into `cmd/<binary-name>`. All
other code is below `lib/`. The remaining scripts in `src/` move into
`util/`.

This commit also introduces a GOPATH at `gopath/` which is used by the
Makefile. I will move more stuff into there in the next commit.
