+++ gopath/src/github.com/holocm/holo
@@ -1,0 +1,1 @@
+../../../..
