+++ lib/holo-files/impl/filebuffer.go
@@ -27,7 +27,7 @@ import (
 	"os"
 	"path/filepath"
 
-	"../common"
+	"github.com/holocm/holo/lib/holo-files/common"
 )
 
 //FileBuffer represents the contents of a file. It is used in holo.Apply() as
