+++ util/distribution-integration/alpm-hook.sh
