+++ util/distribution-integration/alpm.hook
